% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UFOobs.R
\name{UFOobs}
\alias{UFOobs}
\title{Function to pull detailed UFO sightings data by date or date range}
\usage{
UFOobs(year = NA, month = NA, year_end = NA, month_end = NA)
}
\arguments{
\item{year}{year of interest}

\item{month}{month of interest}

\item{year_end}{end year in date range}

\item{year_start}{end month in date range
year and month are required to return UFO sightings data
If no values are provided for year_end or year_start, UFO sightings are returned for a single month

Returned dataframe includes Date, City, State, Shape, Duration, Summary, and date Posted}
}
\description{
Function to pull detailed UFO sightings data by date or date range
}

