% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UFOcount.R
\name{UFOcount}
\alias{UFOcount}
\title{UFOcount()}
\usage{
UFOcount(date = NA)
}
\description{
This function returns a count of the total monthly UFO sightings for a given month & year, or date range.
This defaults to returning a dataframe with all UFO sightings by month for all years available.
date must be in format %Y-%mi.e. 2001-01
For more detailed data on individual observations see UFOobs()
}

